favShops 配列にしてお気に入りの店を返す
